+++ src/main/java/com/elmakers/mine/bukkit/heroes/SkillSelector.java
@@ -1,6 +1,9 @@
 package com.elmakers.mine.bukkit.heroes;
 
-import com.elmakers.mine.bukkit.utility.CompatibilityUtils;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.ClickType;
 import org.bukkit.event.inventory.InventoryAction;
@@ -8,9 +11,7 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import com.elmakers.mine.bukkit.heroes.utilities.CompatibilityUtils;
 
 public class SkillSelector {
     private final HotbarController controller;